refactor(portal): remove portal-components pkg
affects: @fremtind/jkl-button-react, @fremtind/jkl-radio-button-react, @fremtind/portal
