+++ packages/button-react/documentation/buttons.mdx
@@ -5,7 +5,6 @@ react: button-react
 ---
 
 import Example from "./Example";
-import { ComponentExample } from "@fremtind/jkl-portal-components";
 import { PrimaryButton, SecondaryButton, TertiaryButton } from "../src";
 
 <Ingress>
