--- packages/portal-components/README.md
@@ -1,11 +1,0 @@
-# `@fremtind/jkl-portal-components`
-
-> TODO: description
-
-## Usage
-
-```
-const jklPortalComponents = require('@fremtind/jkl-portal-components');
-
-// TODO: DEMONSTRATE API
-```
