--- packages/portal-components/gulpfile.js
@@ -1,3 +1,0 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-require("../../gulpfile")(require("gulp"));
-/* eslint-enable @typescript-eslint/no-var-requires */
