--- packages/portal-components/tsconfig-for-declarations.json
@@ -1,8 +1,0 @@
-{
-    "extends": "../../tsconfig-for-declarations.json",
-    "compilerOptions": {
-        "outDir": "./build",
-        "rootDir": "./src"
-    },
-    "include": ["./src"]
-}
