+++ packages/radio-button-react/package.json
@@ -39,8 +39,7 @@
         "@fremtind/jkl-radio-button": "^1.2.1"
     },
     "devDependencies": {
-        "@fremtind/browserslist-config-jkl": "^0.4.14",
-        "@fremtind/jkl-portal-components": "^0.0.1"
+        "@fremtind/browserslist-config-jkl": "^0.4.14"
     },
     "peerDependencies": {
         "@types/react": "^16.8.17",
