+++ portal/src/components/ComponentExample/ComponentExample.scss
