+++ portal/src/components/Typography/FormatProvider.tsx
@@ -2,7 +2,7 @@ import React, { ReactNode } from "react";
 import { MDXProvider } from "@mdx-js/react";
 import { Link } from "@fremtind/jkl-typography-react";
 import { OrderedList, UnorderedList, ListItem } from "@fremtind/jkl-list-react";
-import { ComponentExample } from "@fremtind/jkl-portal-components";
+import { ComponentExample } from "../ComponentExample";
 
 import {
     PageTitle,
